5_docker_compose_complex/requirements.txt
Flask==2.1.1
flask-redis==0.4.0
